Updated changes from coderabbitai
