+++ src/app/[locale]/servicios/[selectedService]/servicios.module.css
@@ -1,0 +1,21 @@
+.imageCol {
+    text-align: center;
+    padding: 0.9375rem;
+}
+
+.image {
+    border-radius: 0.25rem;
+}
+
+.centerText {
+    text-align: center;
+}
+
+.button {
+    margin-top: 2rem;
+}
+
+.tag {
+    white-space: unset !important;
+    margin-top: 0.625rem;
+}
