--- src/app/[locale]/servicios/[servicios]/servicios.module.css
@@ -1,21 +1,0 @@
-.imageCol {
-  text-align: center;
-  padding: 15px;
-}
-
-.image {
-  border-radius: 4px;
-}
-
-.centerText {
-  text-align: center;
-}
-
-.button {
-  margin-top: 2rem;
-}
-
-.tag {
-  white-space: unset !important;
-  margin-top: 10px;
-}
